Add files via upload
Script to grab information from the computer and append to a file called client_log.txt. If that is not in the directory that it will create a file of the same name. It will append information to the file, grabbing date, time, cpu_load, temperature, and network along with warnings and descriptions